// frontend/src/client/types.gen.ts
export type CreateStaffCommand = {
    role: 'DOCTOR' | 'ADMIN_STAFF';
    firstName?: string;
    lastName?: string;
    email?: string;
    specialization?: string;
    pwz?: string;
    department?: string;
    position?: string;
    workSchedule?: string;
};